requirements.txt
langchain
langgraph
qdrant-client
sentence-transformers
transformers
torch
streamlit
pypdf
pytest
python-dotenv
requests
